+++ RELEASE_NOTES.md
@@ -2,6 +2,8 @@
 
 ## New in git-machete 3.16.1
 
+- fixed: `advance` crashing when the current branch is untracked
+
 ## New in git-machete 3.16.0
 
 - deprecated: `machete.overrideForkPoint.<branch>.whileDescendantOf` is no longer taken into account; it's still written, however, for compatibility reasons